Move reusable flow buttons out
